Move ASG configuration tests to autoscaling_configuration_test.go
